requirements.txt
feedparser >= 5.2
py-trello >= 0.11
six >= 1.11
